cdc: split the initializer from the endpoint
